Add a GDAL source.
This handles more transforms than the mapnik source.
