+++ test/data/landcover_sample_1000.tif.sha512
@@ -1,0 +1,1 @@
+0a1e8c4cf29174b19ddece9a2deb7230a31e819ee78b5ec264feda6356abf60d63763106f1ddad9dd04106d383fd0867bf2db55be0552c30f38ffb530bf72dec
